# tools-workspace/contracts/pdf/openapi.yaml
...
openapi: 3.0.3
info:
  title: EasyToolHub Tool API — PDF domain
  version: 0.1.0
  description: Client-upload PDF microservice for EasyToolHub (temp processing, no permanent storage).
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/pdf/encrypt:
    post:
      summary: Password-protect PDF
      description: >
        `userPassword` and `ownerPassword` must be Base64-encoded UTF-8 strings
        (Angular client encodes; tool-api decodes before encrypting).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, userPassword]
              properties:
                file: { type: string, format: binary }
                userPassword:
                  type: string
                  description: Base64-encoded UTF-8 user password
                ownerPassword:
                  type: string
                  description: Base64-encoded UTF-8 owner password (optional)
                allowPrint: { type: boolean, default: true }
                allowModify: { type: boolean, default: false }
      responses:
        "200":
          description: Encrypted PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
  /api/v1/pdf/decrypt:
    post:
      summary: Remove PDF password
      description: >
        `password` must be Base64-encoded UTF-8 (client encodes; server decodes).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, password]
              properties:
                file: { type: string, format: binary }
                password:
                  type: string
                  description: Base64-encoded UTF-8 password
      responses:
        "200":
          description: Decrypted PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
  /api/v1/pdf/compress:
    post:
      summary: Compress PDF
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                quality: { type: string, enum: [low, medium, high], default: medium }
      responses:
        "200":
          description: Compressed PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
  /api/v1/pdf/merge:
    post:
      summary: Merge PDFs
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  items: { type: string, format: binary }
      responses:
        "200":
          description: Merged PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
  /api/v1/pdf/extract:
    post:
      summary: Extract pages
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, pages]
              properties:
                file: { type: string, format: binary }
                pages: { type: string, description: "e.g. 1,3,5-7" }
      responses:
        "200":
          description: Extracted PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
  /api/v1/pdf/delete-pages:
    post:
      summary: Delete pages
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, pages]
              properties:
                file: { type: string, format: binary }
                pages: { type: string }
      responses:
        "200":
          description: PDF without deleted pages
          content:
            application/pdf:
              schema: { type: string, format: binary }
  /api/v1/pdf/rotate:
    post:
      summary: Rotate pages
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                pages: { type: string }
                degrees: { type: integer, default: 90 }
      responses:
        "200":
          description: Rotated PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
  /api/v1/pdf/reorder:
    post:
      summary: Reorder pages
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, order]
              properties:
                file: { type: string, format: binary }
                order: { type: string, description: "Full page order e.g. 3,1,2" }
      responses:
        "200":
          description: Reordered PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
  /api/v1/pdf/watermark:
    post:
      summary: Add text watermark
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, text]
              properties:
                file: { type: string, format: binary }
                text: { type: string }
                opacity: { type: number, default: 0.25 }
      responses:
        "200":
          description: Watermarked PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
  /api/v1/pdf/page-numbers:
    post:
      summary: Add page numbers
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                startAt: { type: integer, default: 1 }
      responses:
        "200":
          description: Numbered PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
  /api/v1/pdf/metadata:
    post:
      summary: Set metadata
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                title: { type: string }
                author: { type: string }
                subject: { type: string }
                keywords: { type: string }
      responses:
        "200":
          description: Updated PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
  /api/v1/pdf/images-to-pdf:
    post:
      summary: Images to PDF
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  items: { type: string, format: binary }
      responses:
        "200":
          description: PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
  /api/v1/pdf/html-to-pdf:
    post:
      summary: HTML to PDF
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [html]
              properties:
                html: { type: string }
      responses:
        "200":
          description: PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
  /api/v1/pdf/text-to-pdf:
    post:
      summary: Text to PDF
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string }
      responses:
        "200":
          description: PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
  /api/v1/pdf/pdf-to-text:
    post:
      summary: PDF to text
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "200":
          description: Text file
          content:
            text/plain:
              schema: { type: string, format: binary }
  /api/v1/pdf/pdf-to-images:
    post:
      summary: PDF pages to PNG zip
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                dpi: { type: integer, default: 150 }
      responses:
        "200":
          description: ZIP of PNGs
          content:
            application/zip:
              schema: { type: string, format: binary }
  /api/v1/pdf/split:
    post:
      summary: Split PDF into ZIP of page-range parts
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, ranges]
              properties:
                file: { type: string, format: binary }
                ranges:
                  type: string
                  description: Semicolon-separated page specs (e.g. "1-3;4;5-7")
                prefix: { type: string, default: split }
      responses:
        "200":
          description: ZIP of PDFs
          content:
            application/zip:
              schema: { type: string, format: binary }
  /api/v1/pdf/fill-form:
    post:
      summary: Fill AcroForm fields
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, fields]
              properties:
                file: { type: string, format: binary }
                fields:
                  type: string
                  description: JSON object map or array of {name,value}
      responses:
        "200":
          description: Filled PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
  /api/v1/pdf/flatten-form:
    post:
      summary: Flatten AcroForm fields into page content
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "200":
          description: Flattened PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
  /api/v1/pdf/annotate:
    post:
      summary: Stamp annotations/highlights onto PDF pages
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, annotations]
              properties:
                file: { type: string, format: binary }
                annotations:
                  type: string
                  description: JSON array of annotation drafts
      responses:
        "200":
          description: Annotated PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
  /api/v1/pdf/stamp-image:
    post:
      summary: Stamp a PNG/JPEG image onto a PDF page (signatures)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, image, x, y, width, height]
              properties:
                file: { type: string, format: binary }
                image: { type: string, format: binary }
                page: { type: integer, default: 1 }
                x: { type: number }
                y: { type: number }
                width: { type: number }
                height: { type: number }
      responses:
        "200":
          description: Stamped PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
...
  /api/v1/pdf/merge-async:
    post:
      summary: Async large merge (scratch-file cache)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  items: { type: string, format: binary }
      responses:
        "200":
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdfJobStatus'
  /api/v1/pdf/split-async:
    post:
      summary: Async large split
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, ranges]
              properties:
                file: { type: string, format: binary }
                ranges: { type: string, description: 'Semicolon-separated ranges e.g. 1-3;4-6' }
                prefix: { type: string, default: split }
      responses:
        "200":
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdfJobStatus'
...
components:
  schemas:
    PdfJobStatus:
      type: object
      properties:
        jobId: { type: string }
        status:
          type: string
          enum: [QUEUED, RUNNING, COMPLETED, FAILED, EXPIRED]
        operation: { type: string }
        total: { type: integer }
        completed: { type: integer }
        failed: { type: integer }
        progress: { type: number, minimum: 0, maximum: 1 }
        message: { type: string }
        downloadReady: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }